+++ core-ng/src/test/java/core/framework/impl/web/controller/ControllerInspectorTest.java
@@ -1,13 +1,12 @@
 package core.framework.impl.web.controller;
 
+import core.framework.module.LambdaController;
 import core.framework.web.Controller;
 import core.framework.web.Request;
 import core.framework.web.Response;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 /**
  * @author neo
@@ -15,26 +14,26 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 class ControllerInspectorTest {
     @Test
     void methodReference() throws NoSuchMethodException {
-        var inspector = new ControllerInspector(new TestControllers()::get);
-        assertEquals(TestControllers.class, inspector.targetClass);
-        assertEquals(TestControllers.class.getDeclaredMethod("get", Request.class), inspector.targetMethod);
-        assertEquals(TestControllers.class.getCanonicalName() + ".get", inspector.controllerInfo);
+        var inspector = new ControllerInspector((LambdaController) new TestControllers()::get);
+        assertThat(inspector.targetClass).isEqualTo(TestControllers.class);
+        assertThat(inspector.targetMethod).isEqualTo(TestControllers.class.getDeclaredMethod("get", Request.class));
+        assertThat(inspector.controllerInfo).isEqualTo(TestControllers.class.getCanonicalName() + ".get");
     }
 
     @Test
     void lambdaMethod() {
-        var inspector = new ControllerInspector(request -> null);
+        var inspector = new ControllerInspector((LambdaController) request -> null);
         assertThat(inspector.targetClass.getCanonicalName()).startsWith(ControllerInspectorTest.class.getCanonicalName());
-        assertNotNull(inspector.targetMethod);
+        assertThat(inspector.targetMethod).isNotNull();
         assertThat(inspector.controllerInfo).startsWith(ControllerInspectorTest.class.getCanonicalName() + ".");
     }
 
     @Test
     void staticClass() throws NoSuchMethodException {
         var inspector = new ControllerInspector(new TestController());
-        assertEquals(TestController.class, inspector.targetClass);
-        assertEquals(TestController.class.getMethod("execute", Request.class), inspector.targetMethod);
-        assertEquals(TestController.class.getCanonicalName() + ".execute", inspector.controllerInfo);
+        assertThat(inspector.targetClass).isEqualTo(TestController.class);
+        assertThat(inspector.targetMethod).isEqualTo(TestController.class.getMethod("execute", Request.class));
+        assertThat(inspector.controllerInfo).isEqualTo(TestController.class.getCanonicalName() + ".execute");
     }
 
     public static class TestController implements Controller {